include: drop the deprecation markers
They are (IMHO) mostly annoying, rarely helpful. They are also set on a lot of
different kinds of deprecatations: from things that are have no function
anymore to options that just aren't recommended to use because there are
better replacements. Our ABI and API promises are our real guarantees.

They were added with the best intention and I too believed in them at the
time.

The deprecation warnings also trip up the typecheck-gcc system as long as that
includes checks for deprecated options.

The were added in 6967571bf20624bc4cfa6