init_do(): check result
Calls to `Curl_init_do()` did not check on result and missed failures
to properly and completely initialize a transfer request.

The main cause of such an init failure is the need to rewind the READFUNCTION
without a SEEKFUNCTION registered. Check the failure to "rewind" the upload
data immediately make test cases 1576 and friends fail.

refs #17139