meta-hash: hash table with static keys
A variation of our hash tables that uses static `meta_key` structs
to store/lookup entries. Meta keys define the destructor callback
to free values and a "print" function.

Use meta hashes at easy handles and connections to associate values
with an easy handle/connection. Store values get auto-destroyed when
the easy handle is destroyed or reset or when a connection is freed.

DoH and multi event handling are the first users, but protocols
can also make use of this facility. Another use case are the various
STRING_* values in `data->set.str[]` and `data->state.aptr` members.