GHA/windows: limit jobs to 15 minutes
They typically finish (well) within 10 minutes.

A notable exception was vcpkg jobs when a rebuild was triggered.
With losing caches and reducing them to short builds, this is not
an issue at the moment.

The advantage of shorter timeouts is hung/crashed jobs giving back
control earlier for a manual retry.