cmake: honor individual picky option override via `CMAKE_C_FLAGS`
To allow overriding individual picky warning options via for example:
```
-DCMAKE_C_FLAGS=-Wno-xor-used-as-pow
```

Also to sync up with similar `./configure` feature in
`CURL_ADD_COMPILER_WARNINGS`.

It may be useful as a workaround if a specific build combination hits
a picky warning within curl's source code. If such happens, we do
appreciate a report to fix it in curl itself.