cmake: set `BUILDING_LIBCURL` directly for unit test targets
To avoid adding this macro to the global `CURL_DEBUG_MACROS` variable,
which may be used for targets defined after unit tests, and where this
macro may not be necessary.

At the moment unit tests are defined last, so this did not cause any
issue.

Follow-up to 220eda34cdfdcc4316845d2e2bbe5bd8b10305e0 #17259