conn-setup: link dns_entry
The `struct Curl_dns_entry *` passed to setup a connection filter
chain is kept in the setup filters. While the reference in `conn->dns_entry`
is *supposed* to live longer than the connection filters, it seems
better that the connection filters increment and release the reference
counting for that entry.