fix connection match check multi reuse
When re-using a multiplexed connection, we must make sure that
it is in use by transfers from the same multi. Another multi
might run in another thread.

Track in `conn->attached_multi` the use. It is NULL when no
transfers are attached. Otherwise, just check equality on this
pointer for reuse considerations.