Win32/SCHANNEL align --cacert behaviour with OpenSSL and LibreSSL
Win8/Server2012 widened the PKIX chain traversal API to allow certificate
traversal to terminate at an intermediate.

This behaviour (terminate at the fist matching intermediate) is the default
for LibreSSL and OpenSSL (with OpenSSL allowing control via
CURLSSLOPT_NO_PARTIALCHAIN)

This change uses the new API if it is available, and also allows the behaviour
to revert legacy if CURLSSLOPT_NO_PARTIALCHAIN is present.