libssh: Remove free done for sftp_send_aio for successful write
libssh's sftp_aio_wait_write() would free the sftp_aio
and assign NULL to the variable storing it in case of
success and error (not in case of SSH_AGAIN).

Hence freeing sftp_send_aio and assigning it NULL is
not needed in case of successful sftp_aio_wait_write()
due to which this commit removes that code.

Signed-off-by: Eshan Kelkar <eshankelkar@galorithm.com>