follow location: fail early when rewind of input failed
When inspecting a possible follow HTTP request, the result of a
rewind of the upload data was ignored as it was not clear at that
point in time if the request would become a GET.

This initiated the followup, rewound again, which failed again and
terminated the follow up.

This was confusing to users as it was not clear of the follow up
was done or not.

Fix: fail the early rewind when the request is not converted to GET.

refs #17472