ftp: fix teardown of DATA connection in done
When ftp_done() is called to terminate the transfer, it needs to
tear down any open SECONDARY filter chain. The condition on when to
do that was relying on there to be a valid socket. This is not
sufficient as the socket is only set *after* happy eyeballing has
decided on one.

Instead of checking for a valid conn->sock, check if any connection
filter is installed.

refs #17482