SCP/SFTP: avoid busy loop after EAGAIN
The libssh libraries do not reveal if they still have data buffered
from the peer. Only when their buffers are read empty can curl be
sure that it is safe to rely on socket polling.

This change adds detection of EGAIN on receive in the transfer loop
and allows SFTP/SCP transfers to avoid a busy loop in such a case
(which should happen often when CPU exceeds network bandwidth).