Revert "windows: test without fix"
This reverts commit 900a298b24c80dd6fe2d2562a276a11fe923cdd5.

Right, so handling of `~` by cmake has it quirks. On Windows, it's
replaced by a path with backslashes, which the confuses the `cmake_install.cmake`
script, leading to a warning on install with cmake 3.31.x, which
is promoted to an error with cmake 4.0.2.

Perhaps it's not the best to rely on `~` expansion with cmake after
all.