openssl: fix handling of buffered data
`SSL_pending()` only checks if the *current* TLS packet has more
data. There might be more data in SSL's buffers.

`SSL_has_pending()` only checks if there is data in buffers, but
does *not* check if there is a complete TLS packet that can be
decoded.

If we only check the first, we will poll on socket events with
having processed all data and may stall. If we only check the
second, we would busy loop without SSL_read() ever giving something.

Add the flag `connssl->input_pending` that is set on incoming data
in the BIO receive. Clear the flag when encountering a CURLE_AGAIN
on the filters receive (via SSL_read()).

refs #17596