mk-unity: make --srcdir not apply to test source unless embedding
The problem is that we're always passing --srcdir for tests in cmake,
which makes the generated bundle source use absolute paths in it.
This is technically, but ugly, so to avoid it, limit handling --srcdir
for non-test source (aka lib/src with autotools) and when --embed
was also passed.

Another option would have been to pass --srcdir only when also passing
--embed, but the srcdir argument is different for each 4 test group,
and it'd have added a bunch of code for this little niche detail.

The current solution is just two extra condition within an existing
condition. If we ever start generating test bundles for clang-tidy
with autotools (--embed in tests), this may have to be tweaked further.