cmake: fix build if libpsl is not found
There is an option to disable libpsl, which means the library is
optional and not required.

However if libpsl is not explicitly disabled with CURL_USE_LIBPSL
and it is not found, the build will break.

```
CMake Error at C:/Users/xxx/AppData/Local/Programs/CLion/bin/cmake/win/
x64/share/cmake-3.31/Modules/FindPackageHandleStandardArgs.cmake:233
(message):
  Could NOT find Libpsl (missing: LIBPSL_INCLUDE_DIR LIBPSL_LIBRARY)
```

Try to find the package and check if it was found before using it
to unbreak the build, like the other dependencies are already
doing.

Signed-off-by: Alexander Martinz <alex@amartinz.at>