runtests: support memory-limits per test
The idea here is to set limits per test how many allocations and maximum
amount of memory it is allowed to use. This is a means to make sure the
number and total size of allocations are kept in check and don't
mistakenly "blow up".

The <verify> part now supports <limits>, and in this section two limits
can be set for each test (verified in debug builds only):

    Allocations: [number of allocation calls]
    Maximum allocated: [maximum concurrent memory allocated]

If runtests.pl detects that the given limits have been exceeded, it
fails the test case with an error.

Test 1 is the first and so far only test setting these limits.