libssh: Fix handling of sftp_readlink() return value.
The return value of sftp_readlink() was stored
in sshc->readdir_tmp but the code further checked
presence of sshc->readdir_filename before
passing sshc->readdir_tmp to strlen().

Due to this, strlen() would give a segmentation
fault, if sftp_readlink() actually returned NULL
(on error).

This commit changes the handling of the return
value to check presence of sshc->readdir_tmp
(instead of sshc->readdir_filename) before
passing it to strlen()

Signed-off-by: Eshan Kelkar <eshankelkar@galorithm.com>