test1175: fix to run, and fix docs to pass
Fix test1175 to pass the correct source root directory.
Before this patch this caused a silent failure and the tests didn't run.

With this fixed, it detected these issues:
```
CURLE_FUNCTION_NOT_FOUND is not in libcurl-errors.md
CURLE_HTTP_POST_ERROR is not in libcurl-errors.md
CURLE_TELNET_OPTION_SYNTAX is not in libcurl-errors.md
CURLM_CALL_MULTI_SOCKET is not in libcurl-errors.md
```

Fix them by:
- marking `CURLE_FUNCTION_NOT_FOUND` deprecated by 7.53.0
- marking `CURLE_HTTP_POST_ERROR` deprecated by 7.56.0
- marking `CURLE_TELNET_OPTION_SYNTAX` deprecated by 7.78.0
- documenting `CURLM_CALL_MULTI_SOCKET` as a synonym for
  `CURLM_CALL_MULTI_PERFORM`

But, these weren't officially deprecated. It may need more
updates to reflect that in other places, or fix the issues
differently.

Cherry-picked from #17877