config-win32.h: fix default target for VS2012, also bump minimums to XP
When using winbuild or Visual Studio IDE Project Files.

According to CI, the default is Windows 8:
```
-- Found _WIN32_WINNT=0x0602
```
Ref: https://ci.appveyor.com/project/curlorg/curl/builds/51594696/job/elbl0w5n7fmoos2f#L45

This also aligns with the default being Windows 7 for VS2010,
and Windows Vista for VS2008.

Also bump the minimum target to Windows XP, which is what curl supports
and higher than what VS2008 is able to target.