conniection: clarify `transport`
The `transport` to use for a transfer, e.g. TCP/QUIC/UNIX/UDP, is
initially selected by options and protocol used. This is set at
the `struct connectdata` as `transport` member.

During connection establishment, this transport may change due to
Alt-Svc or Happy-Eyeballing. Most common is the switch from TCP
to QUIC.

Rename the connection member to `transport_wanted` and add a way
to query the connection for the transport in use via a new
connection filter query.

The filter query can also be used in the happy eyeballing attempts
when code needs to know which transport is used by the "filter below".
This happens in wolfssl initialization, as one example.