CURLINFO_TLS_SSL_PTR testing
Add tests of CURLINFO_TLS_SSL_PTR and its returned value in test
client 'hx-download'. Use obtained pointer to look up the negotiated
TLS version.

Update manpage of CURLINFO_TLS_SSL_PTR to also describe the
behaviour of wolfSSL similar to OpenSSL. Fix the wolfSSl implementation
for TCP to behave like that. Update the QUIC queries.

Fix rustls `get_internals()` to return the rustls_connection* and
not the address of the pointer.