wincrypt.h is necessary here
The issue with this is that the if_nametoindex auto-detection code would
also need this hack, and it will make those super complicated.

The auto-detection is not normally run with cmake (but necessary to pass
a checkprefill test). With autotools however, it's required to detect
if_nametoindex.

`wincrypt.h` is not safe to include as-is because UWP does not support it,
but in this case this is safe because `if_nametoindex` is not supported by
UWP, and thus `HAVE_IF_NAMETOINDEX` is always undefined.