pollset/getsock, increase limit to 16
Let a pollset keep up to 16 sockets, now that IP happy eyeballing
can have several connect attempts open. Make the return bits from
`getsock()` calls unsigned.

This is a stopgap, as even 16 might not be enough when there is
HTTPS eyeballing in top of IP eyeballing with many A+AAAA entries
that do not respond right away.