replace conn [write]sockfd members by index
The `connectdata` members `sockfd` and `writesockfd` needed to by
either CURL_SOCKET_BAD or a copy of one of `conn->sock[2]`. When
equal to one, that index was used to send/recv/poll the proper
socket or connection filter chain.

Replace those with `send_idx` and `recv_idx` which are either -1,
0 or 1 to indicate which socket/filter to send/receive on.