eliminate request `getheader` bool, reverse `header` default
Deduce that the transfer response expects headers by the protocol
handler implementing `write_resp_hd` callback. This eleminates
the `getheader` parameter in the `Curl_xfer_setup_*()` methods.

Add an implementation to RTSP for `write_resp_hd`, joining the
HTTP protocol in the only handlers having it.

Reverse the default of request's `header` bit that signals that
headers are expected. Default is now FALSE, set to TRUE when
setting up the transfer by presence of `write_resp_hd` in the
protocol handler.