http: introduce CURLOPT_FORBID_RETRY_ON_REUSE
This patch introduces a curlopt to allow application to request
curl to not retry a request upon a broken reused connection.

Currently libcurl transparently retry a failed request over a new
connection if all of the following are true:

- The previous broken connection was a reused connection.
- The previous broken connection was reset.
- Max retries not exceeded.

This behavior relies the follwing assumptions:
- The connection reset might have occurred before the reused
  connection was picked up for reuse.
- The connection reset only becomes visible because libcurl attempted
  to transmit request over such a connection.

The above assumptions do not always hold in the field. The following
sequence will produce identical externally visible behavior:
- Libcurl transmits request over the connection.
- The peer might have received the request, but the connection cuts off
  right before it could send a response (crash or whatever)
- Libcurl receives connection reset shortly after.

In this case libcurl will choose to establish a new TCP connection and
issue the identical request for a second time.
Given a client side internal state machine implementing a protocol
that requires action upon a connection reset, this behavior has chance
to cause the state machine to miss a state transition and affect
correctness.

This behavior used to be hardwired with no way to disable from users
unless the user explicitly requests `CURLOPT_FORBID_REUSE` and
`CURLOPT_FRESH_CONNECT` at the cost of significant overhead.

The new boolean flag `CURLOPT_FORBID_RETRY_ON_REUSE` decouples retries
from connection reuse, allowing a user to explicitly supress this retry
behavior without losing connection reuse benefits. When the option is set,
a connection reset will be returned as an error to the caller.
The caller assumes responsibility to perform retries according to the
application protocol in this case. The flag defaults to FALSE so that
current behavior remains the same.